+++ test/Spec.hs
@@ -6,6 +6,7 @@ import Test.Tasty (defaultMain, testGroup)
 import qualified Spec.Api.V1.Value as Value
 import qualified Spec.Extra.List as List
 import qualified Spec.Extra.Map as Map
+import qualified Spec.Extra.Traversable as Traversable
 
 --------------------------------------------------------------------------------
 
@@ -18,4 +19,5 @@ main = do
             [ testGroup "list utilities" List.tests
             , testGroup "map utilities" Map.tests
             , testGroup "value utilities" Value.tests
+            , testGroup "traversable" Traversable.tests
             ]