+++ CHANGELOG.md
@@ -2,18 +2,34 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
-## 1.1.0 -- 2022-06-24
+## 1.2.0 -- 2022-07-12
 
 ### Added
 
-- Convenience wrapper for `DerivePNewtype`: `DerivePNewtype'`, `DerivePConstantViaNewtype'`
-- Encode product types as lists: `ProductIsData`, `DerivePConstantViaDataList`
-- Encode enum types as integers: `EnumIsData`, `PEnumData` and `DerivePConstantViaEnum`
-- Plutarch helper functions: `pmatchEnum`, `pmatchEnumFromData`
+- `PBoring` type class, representing singleton types.
+- Instances of `PBoring` for various types.
+- `preconst` for `PConst`, which allows safe coercions between different
+  'pretend' types.
+- `PSemiTraversable` instance for `PTagged`.
+- `preplicateA` and `preplicateA_`, allowing for repeated execution of
+  `PApplicative`.
+- `pwhen` and `punless`, mirroring their Haskell counterparts.
+- `preplicate`, mirroring its Haskell counterpart.
 
-### Changed
+### Modified
 
-- Rename `PConstantViaDataList` to `DerivePConstantViaDataList`
+- `PFunctor` now has a `pfconst` method as a back-end for `#$>` and `#<$`. This
+  has a default implementation in terms of `pfmap`.
+- `pvoid` can now replace every location with any `PBoring`, not just `PUnit`.
+- `PTraversable` now has a `ptraverse_` method, which allows us to avoid
+  rebuilding the `PTraversable` if we don't need it anymore. This allows much
+  better folding, for example.
+- `PSemiTraversable` now has a `psemitraverse_` method, with similar benefits to
+  `ptraverse_`.
+- `psemifold`, `psemifoldMap` and `psemifoldComonad` gained a `PSubcategory t a` 
+  constraint, as the 'container' is guaranteed non-empty in such a case.
+- Significant performance improvements for `PTraversable` and `PSemiTraversable`
+  instances.
 
 ## 1.1.0 -- 2022-06-17
 
@@ -31,7 +47,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 #### AssocMap (`Plutarch.Extra.Map.Unsorted`)
 
-- `psort`
+- `psort` 
 - `pkeysEqual`
 - `pmapUnionWith` -> `punionWith`